:tada: use fetch for feedback
